requirements.txt
aiofile
aiohttp[speedups]
appdirs
async_generator
bs4
CairoSVG
cffi
covid
cowpy
eksipy
emoji
faker
git+https://github.com/httplib2/httplib2
git+https://github.com/jmoiron/humanize
git+https://github.com/Joeclinton1/google-images-download
git+https://github.com/LonamiWebs/Telethon/
git+https://github.com/Ncode2014/hentai
git+https://github.com/Ncode2014/pySmartDL
git+https://github.com/Ncode2014/search-engine-parser
git+https://github.com/Ncode2014/youtube_search
git+https://github.com/pawamoy/aria2p
git+https://github.com/rthalley/dnspython
git+https://github.com/TheDeezLoader/pyDeezloader
git+https://github.com/vstinner/hachoir
git+https://github.com/ytdl-org/youtube-dl
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googletrans==3.1.0a0
gTTS
gTTS-token
heroku3
html_telegraph_poster
jikanpy
js2py
lottie
lxml
lyricsgenius
natsort
numpy
oauth2client
pendulum
Pillow
Pillow
psycopg2
psycopg2-binary
pybase64
pydub
pylast
python-barcode
python-dateutil
python-dotenv
pytz
qrcode
rarfile
requests
scipy
speedtest-cli
sqlalchemy<=1.3.23
telegraph
telethon-session-sqlalchemy
telethon-tgcrypto
urbandict
validators
vcsi
wikipedia
wordcloud
